Tweak crawler timings so peers are more likely to be available
