+++ zebrad/src/components/sync/tests/timing.rs
@@ -1,12 +1,16 @@
-use std::sync::{
-    atomic::{AtomicU8, Ordering},
-    Arc,
+use std::{
+    convert::TryInto,
+    sync::{
+        atomic::{AtomicU8, Ordering},
+        Arc,
+    },
 };
 
 use futures::future;
 use tokio::time::{timeout, Duration};
 
-use zebra_chain::parameters::Network;
+use zebra_chain::parameters::{Network, POST_BLOSSOM_POW_TARGET_SPACING};
+use zebra_network::constants::{DEFAULT_CRAWL_NEW_PEER_INTERVAL, HANDSHAKE_TIMEOUT};
 
 use super::super::*;
 use crate::config::ZebradConfig;
@@ -61,9 +65,27 @@ fn ensure_timeouts_consistent() {
 
     // This constraint makes genesis retries more likely to succeed
     assert!(
-        GENESIS_TIMEOUT_RETRY.as_secs() > zebra_network::constants::HANDSHAKE_TIMEOUT.as_secs()
+        GENESIS_TIMEOUT_RETRY.as_secs() > HANDSHAKE_TIMEOUT.as_secs()
             && GENESIS_TIMEOUT_RETRY.as_secs() < BLOCK_DOWNLOAD_TIMEOUT.as_secs(),
         "Genesis retries should wait for new peers, but they shouldn't wait too long"
+    );
+
+    assert!(
+        SYNC_RESTART_DELAY.as_secs()
+            < POST_BLOSSOM_POW_TARGET_SPACING
+                .try_into()
+                .expect("not negative"),
+        "a syncer tip crawl should complete before most new blocks"
+    );
+
+    // The default peer crawler interval should be at least
+    // `HANDSHAKE_TIMEOUT` lower than all other crawler intervals.
+    //
+    // See `DEFAULT_CRAWL_NEW_PEER_INTERVAL` for details.
+    assert!(
+        DEFAULT_CRAWL_NEW_PEER_INTERVAL.as_secs() + HANDSHAKE_TIMEOUT.as_secs()
+            < SYNC_RESTART_DELAY.as_secs(),
+        "an address crawl and peer connections should complete before most syncer tips crawls"
     );
 }
